@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>net48;net6.0-windows;net8.0-windows</TargetFrameworks>
+    <TargetFrameworks>net48;net8.0-windows</TargetFrameworks>
     <!-- we're doing it because tekla is complaning about arm dlls -->
     <!-- we don't really want to support windows arm for now -->
     <PlatformTarget>x64</PlatformTarget> 
