@@ -9,10 +9,10 @@ using Rhino.DocObjects;
 namespace Speckle.Connectors.GrasshopperShared.Parameters;
 
 /// <summary>
-/// The Speckle Property Group Goo is a flat dictionary of (speckle property path, speckle property).
-/// The speckle property path is the concatenated string of all original flattened keys with the property delimiter
+/// The Speckle Property Group Goo is a nested dictionary of (key, speckle property or property group).
+/// The <see cref="Flatten"/> method will use the property delimiter on the keys to flatten the property group into properties.
 /// </summary>
-public partial class SpecklePropertyGroupGoo : GH_Goo<Dictionary<string, SpecklePropertyGoo>>, ISpeckleGoo
+public partial class SpecklePropertyGroupGoo : GH_Goo<Dictionary<string, ISpecklePropertyGoo>>, ISpecklePropertyGoo
 {
   private bool CastFromModelObject(object source)
   {
@@ -22,7 +22,7 @@ public partial class SpecklePropertyGroupGoo : GH_Goo<Dictionary<string, Speckle
         return CastFromModelObject(modelObject.UserText);
 
       case ModelUserText userText:
-        Dictionary<string, SpecklePropertyGoo> dictionary = new();
+        Dictionary<string, ISpecklePropertyGoo> dictionary = new();
         foreach (KeyValuePair<string, string> entry in userText)
         {
           string key = entry.Key;
@@ -45,14 +45,10 @@ public partial class SpecklePropertyGroupGoo : GH_Goo<Dictionary<string, Speckle
     // grasshopper interface types
     if (type == typeof(IGH_ModelContentData))
     {
-      var attributes = new ObjectAttributes();
-      foreach (var entry in Value)
-      {
-        string stringValue = entry.Value.Value?.ToString() ?? "";
-        attributes.SetUserString(entry.Key, stringValue);
-      }
+      ObjectAttributes atts = new();
+      AssignToObjectAttributes(atts);
 
-      var modelObject = new ModelObject(RhinoDoc.ActiveDoc, attributes);
+      ModelObject modelObject = new(RhinoDoc.ActiveDoc, atts);
       target = (T)(object)modelObject;
       return true;
     }
