@@ -1,23 +1,16 @@
 using System.Collections.Concurrent;
 using System.Diagnostics.CodeAnalysis;
 using Autodesk.AutoCAD.DatabaseServices;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
 using Speckle.Connectors.Autocad.HostApp.Extensions;
-using Speckle.Connectors.Autocad.Operations.Send;
 using Speckle.Connectors.Common.Caching;
 using Speckle.Connectors.Common.Cancellation;
-using Speckle.Connectors.Common.Operations;
 using Speckle.Connectors.Common.Threading;
 using Speckle.Connectors.DUI.Bindings;
 using Speckle.Connectors.DUI.Bridge;
-using Speckle.Connectors.DUI.Exceptions;
-using Speckle.Connectors.DUI.Logging;
 using Speckle.Connectors.DUI.Models;
 using Speckle.Connectors.DUI.Models.Card;
 using Speckle.Connectors.DUI.Models.Card.SendFilter;
 using Speckle.Connectors.DUI.Settings;
-using Speckle.Sdk;
 using Speckle.Sdk.Common;
 
 namespace Speckle.Connectors.Autocad.Bindings;
@@ -33,14 +26,11 @@ public abstract class AutocadSendBaseBinding : ISendBinding
   private readonly DocumentModelStore _store;
   private readonly List<ISendFilter> _sendFilters;
   private readonly ICancellationManager _cancellationManager;
-  private readonly IServiceProvider _serviceProvider;
   private readonly ISendConversionCache _sendConversionCache;
-  private readonly IOperationProgressManager _operationProgressManager;
-  private readonly ILogger<AutocadSendBinding> _logger;
-  private readonly ISpeckleApplication _speckleApplication;
   private readonly IThreadContext _threadContext;
   private readonly ITopLevelExceptionHandler _topLevelExceptionHandler;
   private readonly IAppIdleManager _idleManager;
+  private readonly ISendOperationManagerFactory _sendOperationManagerFactory;
 
   /// <summary>
   /// Used internally to aggregate the changed objects' id. Note we're using a concurrent dictionary here as the expiry check method is not thread safe, and this was causing problems. See:
@@ -55,27 +45,21 @@ public abstract class AutocadSendBaseBinding : ISendBinding
     IBrowserBridge parent,
     IEnumerable<ISendFilter> sendFilters,
     ICancellationManager cancellationManager,
-    IServiceProvider serviceProvider,
     ISendConversionCache sendConversionCache,
-    IOperationProgressManager operationProgressManager,
-    ILogger<AutocadSendBinding> logger,
-    ISpeckleApplication speckleApplication,
     IThreadContext threadContext,
     ITopLevelExceptionHandler topLevelExceptionHandler,
-    IAppIdleManager idleManager
+    IAppIdleManager idleManager,
+    ISendOperationManagerFactory sendOperationManagerFactory
   )
   {
     _store = store;
-    _serviceProvider = serviceProvider;
     _cancellationManager = cancellationManager;
     _sendFilters = sendFilters.ToList();
     _sendConversionCache = sendConversionCache;
-    _operationProgressManager = operationProgressManager;
-    _logger = logger;
-    _speckleApplication = speckleApplication;
     _threadContext = threadContext;
     _topLevelExceptionHandler = topLevelExceptionHandler;
     _idleManager = idleManager;
+    _sendOperationManagerFactory = sendOperationManagerFactory;
     Parent = parent;
     Commands = new SendBindingUICommands(parent);
 
@@ -152,55 +136,17 @@ public abstract class AutocadSendBaseBinding : ISendBinding
   {
     try
     {
-      if (_store.GetModelById(modelCardId) is not SenderModelCard modelCard)
-      {
-        // Handle as GLOBAL ERROR at BrowserBridge
-        throw new InvalidOperationException("No publish model card was found.");
-      }
-
-      using var scope = _serviceProvider.CreateScope();
-      InitializeSettings(scope.ServiceProvider);
-
-      using var cancellationItem = _cancellationManager.GetCancellationItem(modelCardId);
-
+      using var manager = _sendOperationManagerFactory.Create();
       // Disable document activation (document creation and document switch)
       // Not disabling results in DUI model card being out of sync with the active document
       // The DocumentActivated event isn't usable probably because it is pushed to back of main thread queue
       Application.DocumentManager.DocumentActivationEnabled = false;
-
-      // Get elements to convert
-      List<AutocadRootObject> autocadObjects = Application.DocumentManager.CurrentDocument.GetObjects(
-        modelCard.SendFilter.NotNull().RefreshObjectIds()
+      await manager.Process(
+        Commands,
+        modelCardId,
+        (sp, card) => InitializeSettings(sp),
+        card => Application.DocumentManager.CurrentDocument.GetObjects(card.SendFilter.NotNull().RefreshObjectIds())
       );
-
-      if (autocadObjects.Count == 0)
-      {
-        // Handle as CARD ERROR in this function
-        throw new SpeckleSendFilterException("No objects were found to convert. Please update your publish filter!");
-      }
-
-      var sendResult = await scope
-        .ServiceProvider.GetRequiredService<SendOperation<AutocadRootObject>>()
-        .Execute(
-          autocadObjects,
-          modelCard.GetSendInfo(_speckleApplication.ApplicationAndVersion),
-          _operationProgressManager.CreateOperationProgressEventHandler(Parent, modelCardId, cancellationItem.Token),
-          cancellationItem.Token
-        );
-
-      await Commands.SetModelSendResult(modelCardId, sendResult.VersionId, sendResult.ConversionResults);
-    }
-    catch (OperationCanceledException)
-    {
-      // SWALLOW -> UI handles it immediately, so we do not need to handle anything for now!
-      // Idea for later -> when cancel called, create promise from UI to solve it later with this catch block.
-      // So have 3 state on UI -> Cancellation clicked -> Cancelling -> Cancelled
-      return;
-    }
-    catch (Exception ex) when (!ex.IsFatal()) // UX reasons - we will report operation exceptions as model card error. We may change this later when we have more exception documentation
-    {
-      _logger.LogModelCardHandledError(ex);
-      await Commands.SetModelError(modelCardId, ex);
     }
     finally
     {
