@@ -1,5 +1,4 @@
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
 using Speckle.Connector.Navisworks.Operations.Send.Filters;
 using Speckle.Connector.Navisworks.Operations.Send.Settings;
 using Speckle.Connector.Navisworks.Services;
@@ -9,7 +8,6 @@ using Speckle.Connectors.Common.Threading;
 using Speckle.Connectors.DUI.Bindings;
 using Speckle.Connectors.DUI.Bridge;
 using Speckle.Connectors.DUI.Exceptions;
-using Speckle.Connectors.DUI.Logging;
 using Speckle.Connectors.DUI.Models;
 using Speckle.Connectors.DUI.Models.Card;
 using Speckle.Connectors.DUI.Models.Card.SendFilter;
@@ -18,7 +16,6 @@ using Speckle.Converter.Navisworks.Settings;
 using Speckle.Converters.Common;
 using Speckle.Sdk;
 using Speckle.Sdk.Common;
-using Speckle.Sdk.Logging;
 
 namespace Speckle.Connector.Navisworks.Bindings;
 
@@ -30,45 +27,36 @@ public class NavisworksSendBinding : ISendBinding
   public SendBindingUICommands Commands { get; }
 
   private readonly DocumentModelStore _store;
-  private readonly IServiceProvider _serviceProvider;
   private readonly ICancellationManager _cancellationManager;
-  private readonly IOperationProgressManager _operationProgressManager;
-  private readonly ILogger<NavisworksSendBinding> _logger;
   private readonly ISpeckleApplication _speckleApplication;
-  private readonly ISdkActivityFactory _activityFactory;
   private readonly INavisworksConversionSettingsFactory _conversionSettingsFactory;
   private readonly ToSpeckleSettingsManagerNavisworks _toSpeckleSettingsManagerNavisworks;
   private readonly IElementSelectionService _selectionService;
   private readonly IThreadContext _threadContext;
+  private readonly ISendOperationManagerFactory _sendOperationManagerFactory;
 
   public NavisworksSendBinding(
     DocumentModelStore store,
     IBrowserBridge parent,
-    IServiceProvider serviceProvider,
     ICancellationManager cancellationManager,
-    IOperationProgressManager operationProgressManager,
-    ILogger<NavisworksSendBinding> logger,
     ISpeckleApplication speckleApplication,
-    ISdkActivityFactory activityFactory,
     INavisworksConversionSettingsFactory conversionSettingsFactory,
     ToSpeckleSettingsManagerNavisworks toSpeckleSettingsManagerNavisworks,
     IElementSelectionService selectionService,
-    IThreadContext threadContext
+    IThreadContext threadContext,
+    ISendOperationManagerFactory sendOperationManagerFactory
   )
   {
     Parent = parent;
     Commands = new SendBindingUICommands(parent);
     _store = store;
-    _serviceProvider = serviceProvider;
     _cancellationManager = cancellationManager;
-    _operationProgressManager = operationProgressManager;
-    _logger = logger;
     _speckleApplication = speckleApplication;
-    _activityFactory = activityFactory;
     _conversionSettingsFactory = conversionSettingsFactory;
     _toSpeckleSettingsManagerNavisworks = toSpeckleSettingsManagerNavisworks;
     _selectionService = selectionService;
     _threadContext = threadContext;
+    _sendOperationManagerFactory = sendOperationManagerFactory;
     SubscribeToNavisworksEvents();
   }
 
@@ -95,60 +83,13 @@ public class NavisworksSendBinding : ISendBinding
 
   private async Task SendInternal(string modelCardId)
   {
-    using var activity = _activityFactory.Start();
-    try
-    {
-      var modelCard = GetModelCard(modelCardId);
-
-      using var scope = _serviceProvider.CreateScope();
-
-      InitializeConverterSettings(scope, modelCard);
-
-      using var cancellationItem = _cancellationManager.GetCancellationItem(modelCardId);
-
-      var progress = _operationProgressManager.CreateOperationProgressEventHandler(
-        Parent,
-        modelCard.ModelCardId.NotNull(),
-        cancellationItem.Token
-      );
-
-      var navisworksModelItems = await GetNavisworksModelItems(modelCard, progress);
-
-      var sendResult = await ExecuteSendOperation(
-        scope,
-        modelCard,
-        navisworksModelItems,
-        progress,
-        cancellationItem.Token
-      );
-
-      await Commands.SetModelSendResult(modelCardId, sendResult.VersionId, sendResult.ConversionResults);
-    }
-    catch (OperationCanceledException)
-    {
-      // SWALLOW -> UI handles it immediately, so we do not need to handle anything for now!
-      // Idea for later -> when cancel called, create promise from UI to solve it later with this catch block.
-      // So have 3 state on UI -> Cancellation clicked -> Cancelling -> Cancelled
-    }
-    catch (Exception ex) when (!ex.IsFatal()) // UX reasons - we will report operation exceptions as model card error. We may change this later when we have more exception documentation
-    {
-      _logger.LogModelCardHandledError(ex);
-      await Commands.SetModelError(modelCardId, ex);
-    }
-    finally
-    {
-      // otherwise the id of the operation persists on the cancellation manager and triggers 'Operations cancelled because of document swap!' message to UI.
-      _cancellationManager.CancelOperation(modelCardId);
-    }
+    using var manager = _sendOperationManagerFactory.Create();
+    await manager.Process(Commands, modelCardId, InitializeConverterSettings, GetNavisworksModelItems);
   }
 
-  private SenderModelCard GetModelCard(string modelCardId) =>
-    _store.GetModelById(modelCardId) as SenderModelCard
-    ?? throw new InvalidOperationException("No publish model card was found.");
-
-  private void InitializeConverterSettings(IServiceScope scope, SenderModelCard modelCard) =>
-    scope
-      .ServiceProvider.GetRequiredService<IConverterSettingsStore<NavisworksConversionSettings>>()
+  private void InitializeConverterSettings(IServiceProvider serviceProvider, SenderModelCard modelCard) =>
+    serviceProvider
+      .GetRequiredService<IConverterSettingsStore<NavisworksConversionSettings>>()
       .Initialize(
         _conversionSettingsFactory.Create(
           originMode: _toSpeckleSettingsManagerNavisworks.GetOriginMode(modelCard),
@@ -159,7 +100,7 @@ public class NavisworksSendBinding : ISendBinding
         )
       );
 
-  private async Task<List<NAV.ModelItem>> GetNavisworksModelItems(
+  private async Task<IReadOnlyList<NAV.ModelItem>> GetNavisworksModelItems(
     SenderModelCard modelCard,
     IProgress<CardProgress> onOperationProgressed
   )
