@@ -11,7 +11,7 @@ using Grasshopper.Rhinoceros.Render;
 
 namespace Speckle.Connectors.GrasshopperShared.Parameters;
 
-public partial class SpeckleObjectWrapperGoo : GH_Goo<SpeckleObjectWrapper>, IGH_PreviewData, ISpeckleGoo
+public partial class SpeckleObjectWrapperGoo : GH_Goo<SpeckleObjectWrapper>, IGH_PreviewData
 {
   public SpeckleObjectWrapperGoo(ModelObject mo)
   {
@@ -114,10 +114,8 @@ public partial class SpeckleObjectWrapperGoo : GH_Goo<SpeckleObjectWrapper>, IGH
         }
       }
 
-      foreach (var kvp in Value.Properties.Value)
-      {
-        atts.SetUserString(kvp.Key, kvp.Value.Value?.ToString() ?? "");
-      }
+      // add props
+      Value.Properties.AssignToObjectAttributes(atts);
 
       target = (T)(object)atts;
       return true;
@@ -165,7 +163,6 @@ public partial class SpeckleObjectWrapperGoo : GH_Goo<SpeckleObjectWrapper>, IGH
         propertyGroup.CastFrom(modelObject.UserText);
 
         // get the object layer
-
         SpeckleCollectionWrapperGoo collWrapperGoo = new();
         SpeckleCollectionWrapper? collWrapper = collWrapperGoo.CastFrom(modelObject.Layer)
           ? collWrapperGoo.Value
@@ -175,9 +172,9 @@ public partial class SpeckleObjectWrapperGoo : GH_Goo<SpeckleObjectWrapper>, IGH
         modelConverted.applicationId = modelObject.Id?.ToString();
         modelConverted[Constants.NAME_PROP] = modelObject.Name.ToString();
         Dictionary<string, object?> propertyDict = new();
-        foreach (var entry in propertyGroup.Value)
+        foreach (var entry in modelObject.UserText)
         {
-          propertyDict.Add(entry.Key, entry.Value.Value);
+          propertyDict.Add(entry.Key, entry.Value);
         }
 
         modelConverted[Constants.PROPERTIES_PROP] = propertyDict;
