@@ -3,11 +3,8 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Speckle.Connectors.Common.Caching;
 using Speckle.Connectors.Common.Cancellation;
-using Speckle.Connectors.Common.Operations;
 using Speckle.Connectors.DUI.Bindings;
 using Speckle.Connectors.DUI.Bridge;
-using Speckle.Connectors.DUI.Exceptions;
-using Speckle.Connectors.DUI.Logging;
 using Speckle.Connectors.DUI.Models;
 using Speckle.Connectors.DUI.Models.Card;
 using Speckle.Connectors.DUI.Models.Card.SendFilter;
@@ -15,9 +12,7 @@ using Speckle.Connectors.DUI.Settings;
 using Speckle.Connectors.TeklaShared.Operations.Send.Settings;
 using Speckle.Converters.Common;
 using Speckle.Converters.TeklaShared;
-using Speckle.Sdk;
 using Speckle.Sdk.Common;
-using Speckle.Sdk.Logging;
 using Tekla.Structures;
 using Tekla.Structures.Model;
 using Task = System.Threading.Tasks.Task;
@@ -31,18 +26,15 @@ public sealed class TeklaSendBinding : ISendBinding
   public IBrowserBridge Parent { get; }
 
   private readonly DocumentModelStore _store;
-  private readonly IServiceProvider _serviceProvider;
   private readonly List<ISendFilter> _sendFilters;
   private readonly ICancellationManager _cancellationManager;
   private readonly ISendConversionCache _sendConversionCache;
-  private readonly IOperationProgressManager _operationProgressManager;
   private readonly ILogger<TeklaSendBinding> _logger;
   private readonly ITeklaConversionSettingsFactory _teklaConversionSettingsFactory;
-  private readonly ISpeckleApplication _speckleApplication;
-  private readonly ISdkActivityFactory _activityFactory;
   private readonly Model _model;
   private readonly ToSpeckleSettingsManager _toSpeckleSettingsManager;
   private readonly Events _events;
+  private readonly ISendOperationManagerFactory _sendOperationManagerFactory;
 
   private ConcurrentDictionary<string, byte> ChangedObjectIds { get; set; } = new();
 
@@ -50,30 +42,24 @@ public sealed class TeklaSendBinding : ISendBinding
     DocumentModelStore store,
     IBrowserBridge parent,
     IEnumerable<ISendFilter> sendFilters,
-    IServiceProvider serviceProvider,
     ICancellationManager cancellationManager,
     ISendConversionCache sendConversionCache,
-    IOperationProgressManager operationProgressManager,
     ILogger<TeklaSendBinding> logger,
     ITeklaConversionSettingsFactory teklaConversionSettingsFactory,
-    ISpeckleApplication speckleApplication,
-    ISdkActivityFactory activityFactory,
-    ToSpeckleSettingsManager toSpeckleSettingsManager
+    ToSpeckleSettingsManager toSpeckleSettingsManager,
+    ISendOperationManagerFactory sendOperationManagerFactory
   )
   {
     _store = store;
-    _serviceProvider = serviceProvider;
     _sendFilters = sendFilters.ToList();
     _cancellationManager = cancellationManager;
     _sendConversionCache = sendConversionCache;
-    _operationProgressManager = operationProgressManager;
     _logger = logger;
     _teklaConversionSettingsFactory = teklaConversionSettingsFactory;
-    _speckleApplication = speckleApplication;
     Parent = parent;
     Commands = new SendBindingUICommands(parent);
-    _activityFactory = activityFactory;
     _toSpeckleSettingsManager = toSpeckleSettingsManager;
+    _sendOperationManagerFactory = sendOperationManagerFactory;
 
     _model = new Model();
     _events = new Events();
@@ -111,55 +97,22 @@ public sealed class TeklaSendBinding : ISendBinding
 
   public async Task Send(string modelCardId)
   {
-    using var activity = _activityFactory.Start();
-
-    try
-    {
-      if (_store.GetModelById(modelCardId) is not SenderModelCard modelCard)
-      {
-        throw new InvalidOperationException("No publish model card was found.");
-      }
-      using var scope = _serviceProvider.CreateScope();
-      scope
-        .ServiceProvider.GetRequiredService<IConverterSettingsStore<TeklaConversionSettings>>()
-        .Initialize(
-          _teklaConversionSettingsFactory.Create(_model, _toSpeckleSettingsManager.GetSendRebarsAsSolid(modelCard))
-        );
-
-      using var cancellationItem = _cancellationManager.GetCancellationItem(modelCardId);
-
-      List<ModelObject> teklaObjects = modelCard
-        .SendFilter.NotNull()
-        .RefreshObjectIds()
-        .Select(id => _model.SelectModelObject(new Identifier(new Guid(id))))
-        .Where(obj => obj != null)
-        .ToList();
-
-      if (teklaObjects.Count == 0)
-      {
-        throw new SpeckleSendFilterException("No objects were found to convert. Please update your publish filter!");
-      }
-
-      var sendResult = await scope
-        .ServiceProvider.GetRequiredService<SendOperation<ModelObject>>()
-        .Execute(
-          teklaObjects,
-          modelCard.GetSendInfo(_speckleApplication.ApplicationAndVersion),
-          _operationProgressManager.CreateOperationProgressEventHandler(Parent, modelCardId, cancellationItem.Token),
-          cancellationItem.Token
-        );
-
-      await Commands.SetModelSendResult(modelCardId, sendResult.VersionId, sendResult.ConversionResults);
-    }
-    catch (OperationCanceledException)
-    {
-      return;
-    }
-    catch (Exception ex) when (!ex.IsFatal())
-    {
-      _logger.LogModelCardHandledError(ex);
-      await Commands.SetModelError(modelCardId, ex);
-    }
+    using var manager = _sendOperationManagerFactory.Create();
+    await manager.Process(
+      Commands,
+      modelCardId,
+      (sp, card) =>
+        sp.GetRequiredService<IConverterSettingsStore<TeklaConversionSettings>>()
+          .Initialize(
+            _teklaConversionSettingsFactory.Create(_model, _toSpeckleSettingsManager.GetSendRebarsAsSolid(card))
+          ),
+      card =>
+        card.SendFilter.NotNull()
+          .RefreshObjectIds()
+          .Select(id => _model.SelectModelObject(new Identifier(new Guid(id))))
+          .Where(obj => obj != null)
+          .ToList()
+    );
   }
 
   public void CancelSend(string modelCardId) => _cancellationManager.CancelOperation(modelCardId);
