@@ -1,6 +1,7 @@
 ﻿using Ara3D.Utils;
 using Speckle.Sdk.Api;
 using Speckle.Sdk.Credentials;
+using Speckle.Sdk.Models.Extensions;
 
 namespace Speckle.Importers.Ifc.Tester2;
 
@@ -16,7 +17,10 @@ public sealed class IfcTester(IClientFactory clientFactory, Importer importer, I
 {
   // Settings, Change these to suit!
   // private readonly ICollection<FilePath> _filePath = [new(@"C:\Users\Jedd\Desktop\GRAPHISOFT_Archicad_Sample_Project-S-Office_v1.0_AC25.ifc")]
-  private readonly IEnumerable<string> _filePaths = Directory.EnumerateFiles(@"C:\Users\Jedd\Desktop\", "*.ifc");
+  private readonly IEnumerable<string> _filePaths = Directory.EnumerateFiles(
+    @"C:\Users\Jedd\Desktop\big files",
+    "*.ifc"
+  );
 
   private readonly Uri _serverUrl = new("https://app.speckle.systems");
   private const string PROJECT_ID = "f3a42bdf24";
@@ -28,7 +32,16 @@ public sealed class IfcTester(IClientFactory clientFactory, Importer importer, I
 
     foreach (var path in _filePaths)
     {
-      await ImportFile(speckleClient, path, cancellationToken);
+      try
+      {
+        await ImportFile(speckleClient, path, cancellationToken);
+      }
+#pragma warning disable CA1031
+      catch (Exception ex)
+#pragma warning restore CA1031
+      {
+        Console.WriteLine(ex.ToFormattedString());
+      }
     }
   }
 
